refactor : 커스텀훅 타입 분리했습니다
